requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
textblob==0.17.1
feedparser==6.0.10
transformers==4.37.2
torch==2.2.0
wordcloud==1.9.3
matplotlib==3.8.2
GoogleNews==1.6.12
pillow==10.2.0
python-dotenv==1.0.1 